Add chroot backup/restore function

+++ src/com/offsec/nethunter/ChrootManagerFragment.java
@@ -99,6 +99,7 @@ public class ChrootManagerFragment extends Fragment {
 
     private static final String URI_MINIMAL = IMAGE_SERVER + FILENAME_MINIMAL;
     private static final String URI_FULL = IMAGE_SERVER + FILENAME_FULL;
+    private static final String FILENAME_BACKUP = "kalifs-backup.tar.gz";
     private final ShellExecuter x = new ShellExecuter();
     /*  Certificate pinning
     openssl s_client -showcerts -connect images.offensive-security.com:443 </dev/null 2>/dev/null|openssl x509 -outform PEM >mycertfile.pem
@@ -291,8 +292,15 @@ public class ChrootManagerFragment extends Fragment {
     private void downloadOrSdcard() {
 
         AlertDialog.Builder adb = new AlertDialog.Builder(getActivity());
-        adb.setTitle("Select Chroot install mode:")
-                .setMessage("Download is the prefered mode. Get the latest chroot from the offsec servers.\n\n Also you can place a custom\nkalifs-[minimal|full].tar.xz in /sdcard\nand skip the download.")
+        adb.setTitle("Select chroot install mode:")
+                .setMessage("Download is the prefered mode. Get the latest chroot from the offsec servers.\n\nYou can place a custom\nkalifs-[minimal|full].tar.xz in /sdcard\nand skip the download.\n\nAlso, You can place a back up kalifs-backup.tar.xz in /sdcard to restore your backup chroot.")
+                .setNegativeButton("Restore from SdCard", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.cancel();
+                        restoreBackup(false);
+                    }
+                })
                 .setNeutralButton("Use SdCard", (dialog, which) -> {
                     dialog.cancel();
                     fullOrMinimal(false);
@@ -301,8 +309,9 @@ public class ChrootManagerFragment extends Fragment {
                     dialog.cancel();
                     fullOrMinimal(true);
                 });
+
         AlertDialog ad = adb.create();
-        ad.setCancelable(false);
+        ad.setCancelable(true);
         ad.show();
     }
 
@@ -345,6 +354,13 @@ public class ChrootManagerFragment extends Fragment {
         ad.show();
     }
 
+    private void restoreBackup(final Boolean shouldDownload) {
+        zipFilePath = nh.SD_PATH + "/" + FILENAME_BACKUP;
+        if (!shouldDownload) {
+            UnziptarTask mytask = new UnziptarTask();
+            mytask.execute();
+            }
+    }
 
     private void addMetaPackages() {
         //for now, we'll hardcode packages in the dialog view.  At some point we'll want to grab them automatically.
@@ -433,15 +449,17 @@ public class ChrootManagerFragment extends Fragment {
 
     private void deleteFile(String filePath) {
         File checkFile = new File(filePath);
-        if (checkFile.exists()) {
-            statusLog(filePath + " found.");
-            statusLog(getActivity().getString(R.string.deletingforroom));
-            if (checkFile.delete()) {
-                statusLog("File deleted.");
-                return;
-            } else {
-                statusLog(getActivity().getString(R.string.problemdeletingoldfile));
-                return;
+        if (!filePath.contains("kalifs-backup.tar.gz")){
+            if (checkFile.exists()) {
+                statusLog(filePath + " found.");
+                statusLog(getActivity().getString(R.string.deletingforroom));
+                if (checkFile.delete()) {
+                    statusLog("File deleted.");
+                    return;
+                } else {
+                    statusLog(getActivity().getString(R.string.problemdeletingoldfile));
+                    return;
+                }
             }
         }
     }
@@ -628,13 +646,18 @@ public class ChrootManagerFragment extends Fragment {
 
                 String fExists = x.RunAsRootOutput("[ -f " + zipFilePath + " ] && echo \"1\" || echo \"0\"");
                 if (fExists.equals("0")) {
-                    Log.d(TAG, "Error: No tar.xz found");
+                    if ( zipFilePath.contains("tar.gz")) Log.d(TAG, "Error: No tar.gz found");
+                    else Log.d(TAG, "Error: No tar.xz found");
                     publishProgress("Error: Missing file: " + zipFilePath + " not found.");
                     return false;
                 }
                 // Decompress, extract, and deploy the .tar.xz to the chroot destination in one step
                 publishProgress(getActivity().getString(R.string.extract_chroot));
-                x.RunAsRootWithException(nh.whichBusybox() + " tar -xJf '" + zipFilePath + "' -C '" + nh.NH_SYSTEM_PATH + "'");
+                if ( zipFilePath.contains("tar.gz")) {
+                    x.RunAsRootOutput(nh.whichBusybox() + " tar -xzf '" + zipFilePath + "' -C '" + nh.NH_SYSTEM_PATH + "'");
+                } else {
+                    x.RunAsRootOutput(nh.whichBusybox() + " tar -xJf '" + zipFilePath + "' -C '" + nh.NH_SYSTEM_PATH + "'");
+                }
             } catch (RuntimeException e) {
                 Log.d(TAG, "Error: ", e);
                 publishProgress("Error: " + e.toString());
